pyproject.toml
[project]
name = "agritech-assistant"
version = "0.1.0"
description = "LLM-powered RAG backend for the Agritech assistant"
authors = [{name = "NTTDATA Agritech Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "google-generativeai>=0.5.2",
    "numpy>=1.26.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agritech_core*"]
